refactor: combineBlockWithPosition함수 파일 위치 이동(순환참조제거)
