+++ tetris/src/pages/play/helper/block.ts
@@ -1,5 +1,4 @@
-import { combineBlockWithPosition } from './table';
-import { Block, BlockType, Position } from './types';
+import { Block, BlockType, Position, Table } from './types';
 import { getRandom } from './utils';
 
 export const BLOCK_MAP: Record<BlockType, Block> = {
@@ -79,4 +78,22 @@ export const getBlockBottomPosition = (block: Block, blockPosition: Position) =>
     });
   });
   return max;
+};
+
+export const combineBlockWithPosition = (block: Block, blockPosition: Position) => {
+  const blockColLength = block.shape.length;
+  const blockRowLength = block.shape[0].length;
+  const table = [...Array(blockPosition.col + blockColLength)].map(() =>
+    Array(blockPosition.row + blockRowLength).fill(null),
+  ) as Table;
+  const addCol = blockPosition.col;
+  const addRow = blockPosition.row;
+  block.shape.forEach((blockShapeCol, col) => {
+    blockShapeCol.forEach((isExistBlock, row) => {
+      if (isExistBlock) {
+        table[col + addCol][row + addRow] = block.type;
+      }
+    });
+  });
+  return table;
 };
