+++ tetris/src/pages/play/helper/table.ts
@@ -1,28 +1,10 @@
 import { produce } from 'immer';
-import { getBlockBottomPosition } from './block';
+import { combineBlockWithPosition, getBlockBottomPosition } from './block';
 import { SETTINGS } from './constants';
 import { Block, Position, Table } from './types';
 
 export const getEmptyTable = (col = SETTINGS.col, row = SETTINGS.row) => {
   return [...Array(col)].map(() => Array(row).fill(null)) as Table;
-};
-
-export const combineBlockWithPosition = (block: Block, blockPosition: Position) => {
-  const blockColLength = block.shape.length;
-  const blockRowLength = block.shape[0].length;
-  const table = [...Array(blockPosition.col + blockColLength)].map(() =>
-    Array(blockPosition.row + blockRowLength).fill(null),
-  ) as Table;
-  const addCol = blockPosition.col;
-  const addRow = blockPosition.row;
-  block.shape.forEach((blockShapeCol, col) => {
-    blockShapeCol.forEach((isExistBlock, row) => {
-      if (isExistBlock) {
-        table[col + addCol][row + addRow] = block.type;
-      }
-    });
-  });
-  return table;
 };
 
 export const combineBlockToTable = (table: Table, block: Block, blockPosition: Position) => {
